services/api-gateway/requirements.txt
fastapi
uvicorn
redis
pydantic
pydantic-settings
psycopg2-binary
boto3
python-multipart
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
numpy<2.0
requests